search: move filter types to filter.ts, remove util.ts
