+++ client/shared/src/search/query/completion.ts
@@ -1,6 +1,5 @@
 import * as Monaco from 'monaco-editor'
 import { escapeRegExp, startCase } from 'lodash'
-import { FILTERS, resolveFilter } from './filters'
 import { Token } from './token'
 import { toMonacoRange } from './decoratedToken'
 import { Omit } from 'utility-types'
@@ -8,7 +7,7 @@ import { Observable } from 'rxjs'
 import { IRepository, IFile, ISymbol, ILanguage, IRepoGroup, ISearchContext } from '../../graphql/schema'
 import { SearchSuggestion } from '../suggestions'
 import { isDefined } from '../../util/types'
-import { FilterType, isNegatableFilter } from './util'
+import { FilterType, isNegatableFilter, resolveFilter, FILTERS } from './filters'
 import { first } from 'rxjs/operators'
 import { SymbolKind } from '../../graphql-operations'
