+++ client/web/src/search/panels/RepositoriesPanel.tsx
@@ -2,8 +2,7 @@ import classNames from 'classnames'
 import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { AuthenticatedUser } from '../../auth'
 import { EventLogResult } from '../backend'
-import { FILTERS } from '../../../../shared/src/search/query/filters'
-import { FilterType } from '../../../../shared/src/search/query/util'
+import { FilterType, FILTERS } from '../../../../shared/src/search/query/filters'
 import { Link } from '../../../../shared/src/components/Link'
 import { LoadingPanelView } from './LoadingPanelView'
 import { Observable } from 'rxjs'
